+++ frontend/src/features/todo/index.tsx
@@ -1,49 +1,94 @@
-import type { OrganizationApiGetListResponse } from "@/client/types.gen";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import { useNavigate } from "@tanstack/react-router";
+import type { TodoGetListResponse } from "@/client/types.gen";
+import { Button } from "@/components/ui/button";
+import { useState } from "react";
+import { TodoCreateDialog } from "./dialog/create-dialog";
 
-export const TodoOrganizationList = ({
-  organizationList,
-}: { organizationList: OrganizationApiGetListResponse }) => {
-  const navigate = useNavigate();
+const statusColor = (status?: string) => {
+  switch (status) {
+    case "NotStarted":
+      return "bg-orange-100 text-orange-800";
+    case "InProgress":
+      return "bg-yellow-100 text-yellow-800";
+    case "Completed":
+      return "bg-green-100 text-green-800";
+    default:
+      return "bg-gray-100 text-gray-800";
+  }
+};
+
+export const TodoList = ({ todoList }: { todoList: TodoGetListResponse }) => {
+  const [isOpenCreateDialog, setIsOpenCreateDialog] = useState<boolean>(false);
   return (
-    <div className="min-h-screen bg-gradient-to-br from-gray-50 to-blue-50 p-6">
-      <div className="max-w-6xl mx-auto">
-        <div className="mb-2">
-          <h1 className="text-3xl font-bold text-gray-900 mb-2">Todoリスト</h1>
-          <p className="text-gray-600">登録されている組織一覧</p>
+    <div className="min-h-screen p-6">
+      <div className="max-w-7xl mx-auto">
+        <div className="w-full flex justify-between mb-2">
+          <h1 className="text-2xl font-bold mb-4 text-gray-900">{todoList?.organizationName}</h1>
+          <Button
+            onClick={() => {
+              setIsOpenCreateDialog(true);
+            }}
+            className="mt-4"
+          >
+            新規作成
+          </Button>
         </div>
-        <div className="bg-white rounded-lg shadow-sm border">
-          <div className="p-4 border-b">
-            <h2 className="text-l font-semibold text-gray-900">組織一覧</h2>
-          </div>
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>組織名</TableHead>
-                <TableHead className="w-[120px]"> </TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {organizationList.map((org) => (
-                <TableRow
-                  key={org.id}
-                  className="cursor-pointer"
-                  onClick={() => navigate({ to: `/${org.id}/todo-list` })}
+        <div className="flex flex-row gap-4">
+          <TodoCulumn
+            list={todoList?.list.filter((todo) => todo.status === "NotStarted") ?? []}
+            status="未着手"
+          />
+          <TodoCulumn
+            list={todoList?.list.filter((todo) => todo.status === "InProgress") ?? []}
+            status="進行中"
+          />
+          <TodoCulumn
+            list={todoList?.list.filter((todo) => todo.status === "Completed") ?? []}
+            status="完了"
+          />
+        </div>
+      </div>
+      <TodoCreateDialog
+        isOpenCreateDialog={isOpenCreateDialog}
+        setIsOpenCreateDialog={setIsOpenCreateDialog}
+        organizationId={todoList.organizationId}
+      />
+    </div>
+  );
+};
+
+const TodoCulumn = ({ list, status }: { list: TodoGetListResponse["list"]; status: string }) => {
+  return (
+    <div className="w-full">
+      <h2 className="text-gray-500 text-l font-bold pl-2 mb-2">{status}</h2>
+      <div className="space-y-2 overflow-y-auto pb-2 w-full h-[calc(100vh-180px)]">
+        {list.length ? (
+          list.map((todo) => (
+            <div
+              key={todo.id}
+              className="flex flex-col md:flex-row md:items-center justify-between bg-white rounded-lg cursor-pointer shadow p-4 border hover:bg-gray-100 transition"
+            >
+              <div className="flex-1 min-w-0">
+                <div className="flex items-center gap-2">
+                  <span className="text-lg font-semibold text-gray-800">{todo.title}</span>
+                </div>
+                <div className="mt-1 text-gray-600 text-sm break-words">
+                  {todo.description ?? <span className="text-gray-400">説明なし</span>}
+                </div>
+              </div>
+              <div className="mt-3 md:mt-0 md:ml-6 flex-shrink-0">
+                <span
+                  className={`inline-block px-3 py-1 rounded-full text-xs font-medium ${statusColor(
+                    todo.status
+                  )}`}
                 >
-                  <TableCell className="font-medium">{org.name}</TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </div>
+                  {status}
+                </span>
+              </div>
+            </div>
+          ))
+        ) : (
+          <div className="text-gray-500 text-center py-10">タスクはありません</div>
+        )}
       </div>
     </div>
   );
