+++ frontend/src/features/users/index.tsx
@@ -13,9 +13,9 @@ import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip
 import { useAuthStore } from "@/store/auth-store";
 import { Pencil, Trash2 } from "lucide-react";
 import { useState } from "react";
-import { UserCreateDialog } from "./dialog/create-user-dialog";
-import { UserDeleteDialog } from "./dialog/delete-user-dialog";
-import { UserUpdateDialog } from "./dialog/update-user-dialog";
+import { UserCreateDialog } from "./dialog/create-dialog";
+import { UserDeleteDialog } from "./dialog/delete-dialog";
+import { UserUpdateDialog } from "./dialog/update-dialog";
 
 export const OrganizationProfile = ({
   organization,
