+++ frontend/src/routes/_protected/organizations/$organizationId/todos/index.tsx
@@ -1,10 +1,10 @@
 import { todoGetListOptions } from "@/client/@tanstack/react-query.gen";
-import { TodoList } from "@/features/todo/todo";
+import { TodoList } from "@/features/todo";
 import { queryClient } from "@/lib/query-client";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { createFileRoute, useParams } from "@tanstack/react-router";
 
-export const Route = createFileRoute("/_protected/$organizationId/todo-list/")({
+export const Route = createFileRoute("/_protected/organizations/$organizationId/todos/")({
   loader: async ({ params }) => {
     const { organizationId } = params;
     await queryClient.ensureQueryData(
@@ -16,7 +16,9 @@ export const Route = createFileRoute("/_protected/$organizationId/todo-list/")({
     );
   },
   component: () => {
-    const { organizationId } = useParams({ from: "/_protected/$organizationId/todo-list/" });
+    const { organizationId } = useParams({
+      from: "/_protected/organizations/$organizationId/todos/",
+    });
     const { data: todoList } = useSuspenseQuery(
       todoGetListOptions({
         path: {
